+++ axelor-bank-payment/src/main/java/com/axelor/apps/bankpayment/service/batch/BatchBankPaymentServiceImpl.java
@@ -287,7 +287,8 @@ public class BatchBankPaymentServiceImpl implements BatchBankPaymentService {
                   currency,
                   bankOrderDate,
                   receiverReference,
-                  receiverLabel);
+                  receiverLabel,
+                  paymentScheduleLine);
           bankOrder.addBankOrderLineListItem(bankOrderLine);
         }