+++ axelor-base/src/main/java/com/axelor/apps/base/service/PartnerService.java
@@ -365,8 +365,8 @@ public class PartnerService {
   public void convertToIndividualPartner(Partner partner) {
     partner.setIsContact(false);
     partner.setPartnerTypeSelect(PartnerRepository.PARTNER_TYPE_INDIVIDUAL);
-    addPartnerAddress(partner, partner.getContactAddress(), true, false, false);
-    partner.setContactAddress(null);
+    addPartnerAddress(partner, partner.getMainAddress(), true, false, false);
+    partner.setMainAddress(null);
   }
 
   /**