+++ axelor-production/src/main/java/com/axelor/apps/production/web/ProductionOrderSaleOrderController.java
@@ -21,7 +21,9 @@ import com.axelor.apps.production.db.ProductionOrder;
 import com.axelor.apps.production.exceptions.IExceptionMessage;
 import com.axelor.apps.production.service.ProductionOrderSaleOrderService;
 import com.axelor.apps.sale.db.SaleOrder;
+import com.axelor.apps.sale.db.repo.SaleOrderRepository;
 import com.axelor.exception.AxelorException;
+import com.axelor.exception.service.TraceBackService;
 import com.axelor.i18n.I18n;
 import com.axelor.meta.schema.actions.ActionView;
 import com.axelor.rpc.ActionRequest;
@@ -35,24 +37,42 @@ import java.util.List;
 public class ProductionOrderSaleOrderController {
 
   @Inject ProductionOrderSaleOrderService productionOrderSaleOrderService;
+  @Inject SaleOrderRepository saleOrderRepository;
 
   public void createProductionOrders(ActionRequest request, ActionResponse response)
       throws AxelorException {
 
-    SaleOrder saleOrder = request.getContext().asType(SaleOrder.class);
+    try {
 
-    List<Long> productionOrderIdList =
-        productionOrderSaleOrderService.generateProductionOrder(saleOrder);
-    if (!productionOrderIdList.isEmpty()) {
-      response.setView(
-          ActionView.define(I18n.get("Production order"))
-              .model(ProductionOrder.class.getName())
-              .add("grid", "production-order-grid")
-              .add("form", "production-order-form")
-              .domain("self.id in (" + Joiner.on(",").join(productionOrderIdList) + ")")
-              .map());
-    } else {
-      response.setFlash(I18n.get(IExceptionMessage.PRODUCTION_ORDER_NO_GENERATION));
+      SaleOrder saleOrder = request.getContext().asType(SaleOrder.class);
+      saleOrder = saleOrderRepository.find(saleOrder.getId());
+
+      List<Long> productionOrderIdList =
+          productionOrderSaleOrderService.generateProductionOrder(saleOrder);
+
+      if (productionOrderIdList != null && productionOrderIdList.size() == 1) {
+        response.setView(
+            ActionView.define(I18n.get("Production order"))
+                .model(ProductionOrder.class.getName())
+                .add("form", "production-order-form")
+                .add("grid", "production-order-grid")
+                .param("forceEdit", "true")
+                .context("_showRecord", String.valueOf(productionOrderIdList.get(0)))
+                .map());
+      } else if (productionOrderIdList != null && productionOrderIdList.size() > 1) {
+        response.setView(
+            ActionView.define(I18n.get("Production order"))
+                .model(ProductionOrder.class.getName())
+                .add("grid", "production-order-grid")
+                .add("form", "production-order-form")
+                .domain("self.id in (" + Joiner.on(",").join(productionOrderIdList) + ")")
+                .map());
+      } else {
+        response.setFlash(I18n.get(IExceptionMessage.PRODUCTION_ORDER_NO_GENERATION));
+      }
+
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
     }
   }
 }