+++ axelor-stock/src/main/java/com/axelor/apps/stock/service/InventoryService.java
@@ -60,6 +60,8 @@ import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.time.LocalDate;
 import java.time.ZoneOffset;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.Comparator;
@@ -575,24 +577,43 @@ public class InventoryService {
     List<String[]> list = new ArrayList<>();
 
     for (InventoryLine inventoryLine : inventory.getInventoryLineList()) {
-      String[] item = new String[7];
+      String[] item = new String[9];
       String realQty = "";
 
       item[0] = (inventoryLine.getProduct() == null) ? "" : inventoryLine.getProduct().getName();
       item[1] = (inventoryLine.getProduct() == null) ? "" : inventoryLine.getProduct().getCode();
-      item[2] = (inventoryLine.getRack() == null) ? "" : inventoryLine.getRack();
-      item[3] =
+      item[2] =
+          (inventoryLine.getProduct() == null)
+              ? ""
+              : ((inventoryLine.getProduct().getProductCategory() == null)
+                  ? ""
+                  : inventoryLine.getProduct().getProductCategory().getName());
+      item[3] = (inventoryLine.getRack() == null) ? "" : inventoryLine.getRack();
+      item[4] =
           (inventoryLine.getTrackingNumber() == null)
               ? ""
               : inventoryLine.getTrackingNumber().getTrackingNumberSeq();
-      item[4] = inventoryLine.getCurrentQty().toString();
+      item[5] = inventoryLine.getCurrentQty().toString();
       if (inventoryLine.getRealQty() != null
           && inventory.getStatusSelect() != InventoryRepository.STATUS_DRAFT
           && inventory.getStatusSelect() != InventoryRepository.STATUS_PLANNED) {
         realQty = inventoryLine.getRealQty().toString();
       }
-      item[5] = realQty;
-      item[6] = (inventoryLine.getDescription() == null) ? "" : inventoryLine.getDescription();
+      item[6] = realQty;
+      item[7] = (inventoryLine.getDescription() == null) ? "" : inventoryLine.getDescription();
+
+      String lastInventoryDateTString = "";
+      StockLocationLine stockLocationLine =
+          stockLocationLineService.getStockLocationLine(
+              inventory.getStockLocation(), inventoryLine.getProduct());
+      if (stockLocationLine != null) {
+        ZonedDateTime lastInventoryDateT = stockLocationLine.getLastInventoryDateT();
+        lastInventoryDateTString =
+            lastInventoryDateT == null
+                ? ""
+                : lastInventoryDateT.format(DateTimeFormatter.ofPattern("dd/MM/yyyy"));
+      }
+      item[8] = lastInventoryDateTString;
       list.add(item);
     }
 
@@ -615,11 +636,13 @@ public class InventoryService {
     String[] headers = {
       I18n.get("Product Name"),
       I18n.get("Product Code"),
+      I18n.get("Product category"),
       I18n.get("Rack"),
       I18n.get("Tracking Number"),
       I18n.get("Current Quantity"),
       I18n.get("Real Quantity"),
-      I18n.get("Description")
+      I18n.get("Description"),
+      I18n.get("Last Inventory date")
     };
     CsvTool.csvWriter(filePath, fileName, ';', '"', headers, list);